feat: implement "workspaces" feature
Introduce new 'crates' field to support
cargo workspaces - multiproject repositories.

If non-empty 'crates' configuration value
is provided, then the plugin will bumb the
version of the crate in each specified
subfolder.

Note that there is a single new version
value for all crates. Note also that a crate
name and subfolder in which it located should
be the same.

Implementing fetching the proper crate's id
from its manifest otherwise requires parsing
TOML file. `cargo read-manifest` prints manifest
as JSON, but it is deprecated.

The other possibility is to use `cd <crate>`,
`cargo pkgid`, producing something like
"path+file:///<absolute-path>#<crate-name>@1.0.0".
This requires parsing URI, extracting "hash"
value after # and splitting it on @.

For the above reasons this implementation is
basic and may be not suitable for all cases.